+++ src/websockets/exceptions.py
@@ -394,11 +394,11 @@ class PayloadTooBig(WebSocketException):
         self,
         size_or_message: int | None | str,
         max_size: int | None = None,
-        cur_size: int | None = None,
+        current_size: int | None = None,
     ) -> None:
         if isinstance(size_or_message, str):
             assert max_size is None
-            assert cur_size is None
+            assert current_size is None
             warnings.warn(  # deprecated in 14.0 - 2024-11-09
                 "PayloadTooBig(message) is deprecated; "
                 "change to PayloadTooBig(size, max_size)",
@@ -410,8 +410,8 @@ class PayloadTooBig(WebSocketException):
             self.size: int | None = size_or_message
             assert max_size is not None
             self.max_size: int = max_size
-            self.cur_size: int | None = None
-            self.set_current_size(cur_size)
+            self.current_size: int | None = None
+            self.set_current_size(current_size)
 
     def __str__(self) -> str:
         if self.message is not None:
@@ -420,16 +420,16 @@ class PayloadTooBig(WebSocketException):
             message = "frame "
             if self.size is not None:
                 message += f"with {self.size} bytes "
-            if self.cur_size is not None:
-                message += f"after reading {self.cur_size} bytes "
+            if self.current_size is not None:
+                message += f"after reading {self.current_size} bytes "
             message += f"exceeds limit of {self.max_size} bytes"
             return message
 
-    def set_current_size(self, cur_size: int | None) -> None:
-        assert self.cur_size is None
-        if cur_size is not None:
-            self.max_size += cur_size
-            self.cur_size = cur_size
+    def set_current_size(self, current_size: int | None) -> None:
+        assert self.current_size is None
+        if current_size is not None:
+            self.max_size += current_size
+            self.current_size = current_size
 
 
 class InvalidState(WebSocketException, AssertionError):